backend/requirements.txt
Flask>=3.0.0
flask-cors
google-generativeai>=0.8.3
python-dotenv
scikit-learn
requests
pandas
SQLAlchemy
psycopg2-binary
ijson
xgboost